use chained deobfuscation
